refactor: 좌우 패널 전역 IDE 셸로 공통화
